Update user dto-s

+++ Lendship.Backend/DTO/UserDetailsDto.cs
@@ -81,7 +81,7 @@ namespace Lendship.Backend.DTO
         /// </summary>
 
         [DataMember(Name = "evaluationAsAdvertiser")]
-        public int? EvaluationAsAdvertiser { get; set; }
+        public decimal? EvaluationAsAdvertiser { get; set; }
 
         /// <summary>
         /// Gets or Sets EvaluationAsAdvertiserCount
@@ -95,7 +95,7 @@ namespace Lendship.Backend.DTO
         /// </summary>
 
         [DataMember(Name = "evaluationAsLender")]
-        public int? EvaluationAsLender { get; set; }
+        public decimal? EvaluationAsLender { get; set; }
 
         /// <summary>
         /// Gets or Sets EvaluationAsLenderCount
